Excludes comments.

+++ test.sh
@@ -5,7 +5,9 @@ php_versions=("php83" "php82" "php81")
 
 for version in "${php_versions[@]}"; do
   echo "Setting up environment..."
-  rm composer.lock
+  if [ -f "composer.lock" ]; then
+    rm composer.lock
+  fi
   docker compose run --rm "$version"composer composer install
   echo "Running tests on $version..."
   if ! docker compose run --rm "$version" vendor/bin/phpunit; then
